# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: Failed to parse user data
        type: string
      message:
        example: Something went wrong
        type: string
    type: object
...
  dto.MediaFileListItemResponse:
    properties:
      created_at:
        type: string
      download_url:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  dto.MediaFileResponse:
    properties:
      created_at:
        type: string
      download_url:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.MediaFilesListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.MediaFileListItemResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: This API supports both Web and admin (CMS) operations, including authentication,
    page management.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /cms/media-files:
    get:
      description: Retrieves a paginated list of media files, with optional search
        by name and sorting.
      parameters:
      - description: Search term for file name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      - description: 'Sort by ''name'' or ''created_at'' (default: ''created_at'')'
        enum:
        - name
        - created_at
        in: query
        name: sortBy
        type: string
      - description: 'Sort order ''asc'' or ''desc'' (default: ''desc'')'
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaFilesListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Media Files
      tags:
      - CMS - Media Files
    post:
      consumes:
      - multipart/form-data
      description: Uploads a new media file (image, pdf, svg).
      parameters:
      - description: The file to upload (jpg, png, webp, svg, pdf)
        in: formData
        name: file
        required: true
        type: file
      - description: Optional sub-directory path for the file
        in: formData
        name: path
        type: string
      - description: 'Optional. If true, replaces existing file with the same name
          and path. Default: false.'
        in: formData
        name: replace
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            $ref: '#/definitions/dto.MediaFileResponse'
        "400":
          description: Bad Request (e.g., no file, invalid file type, invalid path)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Upload Media File
      tags:
      - CMS - Media Files
  /cms/media-files/{id}:
    delete:
      description: Deletes a media file by its UUID from both the database and the
        disk.
      parameters:
      - description: Media File ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content - File deleted successfully
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Media file not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error (e.g., failed to delete file from disk
            or DB)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Media File
      tags:
      - CMS - Media Files
    get:
      description: Retrieves details for a specific media file by its UUID.
      parameters:
      - description: Media File ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaFileResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Media file not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Media File by ID
      tags:
      - CMS - Media Files
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"